refactor: move stackset.IsEmptyErr to cloudformation.IsEmptyErr
